+++ app.territoiresentransitions.react/src/app/pages/collectivite/PlanActions/PlanActions.tsx
@@ -1,142 +1,17 @@
-import {
-  useAllFiches,
-  useAllStorables,
-  useEpciId,
-  useStorable,
-} from 'core-logic/hooks';
+import {useAllFiches, useEpciId, useStorable} from 'core-logic/hooks';
 import {PlanActionStorable} from 'storables/PlanActionStorable';
 import {planActionStore} from 'core-logic/api/hybridStores';
 import {useParams} from 'react-router-dom';
-import {Chip} from '@material-ui/core';
-import {
-  Categorie,
-  PlanActionStructure,
-  PlanActionTyped,
-} from 'types/PlanActionTypedInterface';
-import {FicheAction} from 'generated/models/fiche_action';
-import {compareIndexes} from 'utils';
+import {Categorie, PlanActionTyped} from 'types/PlanActionTypedInterface';
 import {FicheCard} from 'app/pages/collectivite/PlanActions/FicheCard';
 import {Spacer} from 'ui/shared';
-
-interface Categorized {
-  fiches: FicheAction[];
-  categorie: Categorie;
-}
-
-const defaultCategorie: Categorie = {
-  nom: 'Sans catégorie',
-  uid: '',
-};
-
-function categorizeAndSortFiches(
-  allFiches: FicheAction[],
-  plan: PlanActionTyped
-): Categorized[] {
-  // step 1: sort categories
-  const categories: Categorie[] = [...plan.categories, defaultCategorie];
-  const fichesByCategory = (plan as PlanActionStructure).fiches_by_category;
-  categories.sort((a, b) => compareIndexes(a.nom, b.nom));
-  return categories.map((categorie: Categorie) => {
-    // step 2: find fiches
-    const fiches: FicheAction[] = [];
-    for (const {fiche_uid} of fichesByCategory.filter(
-      fc => fc.category_uid === categorie.uid
-    )) {
-      const fiche = allFiches.find(f => f.uid === fiche_uid);
-      if (fiche) fiches.push(fiche);
-    }
-    // step 3: sort fiches
-    fiches.sort((a, b) => compareIndexes(a.titre, b.titre));
-    fiches.sort((a, b) => compareIndexes(a.custom_id, b.custom_id));
-    return {
-      categorie: categorie,
-      fiches: fiches,
-    };
-  });
-}
-
-interface CategorizedNode {
-  fiches: FicheAction[];
-  categorie: Categorie;
-  children: CategorizedNode[];
-}
-
-function nestCategorized(categorized: Categorized[]): CategorizedNode[] {
-  // Tree
-  const root: CategorizedNode[] = categorized
-    .filter(c => !c.categorie.parent)
-    .map(c => {
-      return {
-        fiches: c.fiches,
-        categorie: c.categorie,
-        children: [],
-      };
-    });
-
-  function search(
-    nodes: CategorizedNode[],
-    uid: string
-  ): CategorizedNode | null {
-    for (const node of nodes) {
-      if (node.categorie.uid === uid) return node;
-      const found = search(node.children, uid);
-      if (found) return found;
-    }
-    return null;
-  }
-
-  // Consume categorized to put them in the tree.
-  for (const c of categorized) {
-    if (!c.categorie.parent) continue;
-    const node = {
-      fiches: c.fiches,
-      categorie: c.categorie,
-      children: [],
-    };
-    const parent = search(root, node.categorie.parent!);
-    if (parent) parent.children.push(node);
-  }
-  return root;
-}
-
-function PlanNavChip(props: {
-  epciId: string;
-  planUid: string;
-  planNom: string;
-  active: boolean;
-}) {
-  return (
-    <div className="mr-2">
-      <Chip
-        label={props.planNom}
-        component="a"
-        href={`/collectivite/${props.epciId}/plan_action/${props.planUid}`}
-        color={props.active ? 'primary' : 'default'}
-        clickable
-      />
-    </div>
-  );
-}
-
-function PlanNav() {
-  const {epciId, planUid} = useParams<{epciId: string; planUid: string}>();
-  const plans = useAllStorables<PlanActionStorable>(planActionStore);
-  plans.sort((a, b) => a.nom.localeCompare(b.nom));
-
-  return (
-    <nav className="flex flex-row">
-      {plans.map(plan => (
-        <PlanNavChip
-          epciId={epciId}
-          planUid={plan.uid}
-          planNom={plan.nom}
-          active={plan.uid === planUid}
-          key={plan.uid}
-        />
-      ))}
-    </nav>
-  );
-}
+import {PlanNav} from 'app/pages/collectivite/PlanActions/PlanNav';
+import {
+  categorizeAndSortFiches,
+  CategorizedNode,
+  defaultCategorie,
+  nestCategorized,
+} from 'app/pages/collectivite/PlanActions/sorting';
 
 function ModificationDialogButton(props: {categorie: Categorie}) {
   return <div>todo modif</div>;
@@ -158,7 +33,7 @@ function CategoryTitle(props: {categorie: Categorie; editable: boolean}) {
   );
 }
 
-function NestedCategory(props: {nodes: CategorizedNode[]}) {
+function CategoryLevel(props: {nodes: CategorizedNode[]}) {
   return (
     <>
       {props.nodes.map(node => {
@@ -177,7 +52,7 @@ function NestedCategory(props: {nodes: CategorizedNode[]}) {
             })}
             {node.children && (
               <div className="ml-5">
-                <NestedCategory nodes={node.children} />
+                <CategoryLevel nodes={node.children} />
               </div>
             )}
           </div>
@@ -193,7 +68,7 @@ function Plan(props: {plan: PlanActionTyped}) {
   const sorted = categorizeAndSortFiches(fiches, props.plan);
   const nested = nestCategorized(sorted);
 
-  return <NestedCategory nodes={nested} />;
+  return <CategoryLevel nodes={nested} />;
 }
 
 const PlanActions = function () {
